backend/pyproject.toml
[build-system]
requires = ["uv_build>=0.7.21,<0.8.0"]
build-backend = "uv_build"

[project]
name = "pennylane-support"
version = "0.1.0"
description = "PennyLane Support API"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Dave Mariano"},
]
dependencies = [
    "fastapi[all]>=0.95.0",
    "uvicorn[standard]>=0.21.1",
    "python-dotenv>=1.0.0",
    "sqlmodel>=0.0.24",
    "psycopg2-binary>=2.9.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.1",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
]
